# client/api/swagger.yaml
openapi: 3.0.0
info:
  title: JumpCloud Project API
  description: This is a simple API for JumpCloud Project
  contact:
    email: petebirk@memyselfandi.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/pbirk007/JCProject/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://localhost:443/api/v1
  description: Local server
...
paths:
...
  /shutdown:
    post:
      tags:
      - admins
      summary: Shuts down the system after all requests processed.
      description: |
        This will reject any new requests from coming in immediately and will wait for pending requests to finish before exiting.
      operationId: shutdown
      responses:
        "200":
          description: Successfully initiated shutdown.
        "401":
          description: The user does not have authority.
        "405":
          description: Method is not allowed.
...